Implemented uploading to .tmp file first.

+++ app/src/main/java/dev/nikita_chernikov/lab5/DownloadService.kt
@@ -30,10 +30,13 @@ class DownloadService : Service() {
     }
 
     private suspend fun downloadFile(userInput: String) {
-        val tempFile = File(journalDir, "$userInput.tmp")
+        val finalFile = File(journalDir, "rfc$userInput.pdf")
+        val tempFile = File(journalDir, "rfc$userInput.pdf.tmp")
+        var successful = false
+
         try {
             tempFile.createNewFile()
-            DownloadStatusManager.postUpdate()
+            DownloadStatusManager.postUpdate() // Notify UI that download has started
 
             val finalUrl = if (userInput.startsWith("http")) userInput else "https://www.rfc-editor.org/rfc/rfc$userInput.pdf"
             val url = URL(finalUrl)
@@ -41,12 +44,15 @@ class DownloadService : Service() {
             connection.connect()
 
             if (connection.responseCode == HttpURLConnection.HTTP_OK) {
-                val finalFile = File(journalDir, "rfc$userInput.pdf")
-                FileOutputStream(finalFile).use { output ->
+                // Download TO the temp file
+                FileOutputStream(tempFile).use { output ->
                     connection.inputStream.use { input ->
                         input.copyTo(output)
                     }
                 }
+                // If download is complete, rename the file
+                tempFile.renameTo(finalFile)
+                successful = true
             } else if (connection.responseCode == HttpURLConnection.HTTP_NOT_FOUND) {
                 CoroutineScope(Dispatchers.Main).launch {
                     Toast.makeText(applicationContext, "File rfc$userInput.pdf not found (404)", Toast.LENGTH_LONG).show()
@@ -54,11 +60,13 @@ class DownloadService : Service() {
             }
         } catch (_: Exception) {
             CoroutineScope(Dispatchers.Main).launch {
-                Toast.makeText(applicationContext, "Download failed for rfc$userInput.pdf: unexpected error.", Toast.LENGTH_LONG).show()
+                Toast.makeText(applicationContext, "Download failed for rfc$userInput.pdf.", Toast.LENGTH_LONG).show()
             }
         } finally {
-            tempFile.delete()
-            DownloadStatusManager.postUpdate()
+            if (!successful) {
+                tempFile.delete()
+            }
+            DownloadStatusManager.postUpdate() // Notify UI that the process is finished (either success or fail)
         }
     }
